Got the score adding up when you buy items
The score is the amount of money saved by the discount.
Next up.. making a small canvas that shows up over items that shows how much money they are and at what discount

+++ Assets/Scripts/DataContainers/Discount.cs
@@ -4,6 +4,6 @@
 public class Discount : Weighted
 {
     public Sprite sign;
-    public int minDiscount;
-    public int maxDiscount;
+    public float minDiscount;
+    public float maxDiscount;
 }
